Minor logging changes to show app activity.
