+++ timeseries/scala/timeseries-weather/src/main/resources/application.conf
@@ -19,10 +19,7 @@ weather {
 
   # The raw weather data year range to iterate, within 2000
   # right now not downloading, only using 1 local in /${data.dir}
-  data {
-    load.path = "./data/load"
-    file.extension = ".csv.gz"
-  }
+  data.path = "./data/load"
 
   kafka-dispatcher {
     # Dispatcher is the name of the event-based dispatcher
