+++ timeseries/scala/timeseries-weather/src/main/resources/log4j.properties
@@ -6,12 +6,13 @@ log4j.appender.console.layout=org.apache.log4j.PatternLayout
 log4j.appender.console.layout.ConversionPattern=%d{yy/MM/dd HH:mm:ss} %p %c{1}: %m%n
 
 # Settings to quiet third party logs that are too verbose
+log4j.logger.org.apache.thrift.server.TNonblockingServer=ERROR
+log4j.logger.org.apache.cassandra.db.SystemKeyspace=ERROR
 log4j.logger.org.eclipse.jetty=WARN
 log4j.logger.org.apache.spark.repl.SparkIMain$exprTyper=INFO
 log4j.logger.org.apache.spark.repl.SparkILoop$SparkILoopInterpreter=INFO
 log4j.logger.org.apache.hadoop.io.compress.snappy.LoadSnappy=ERROR
 log4j.logger.org.apache.hadoop.util.NativeCodeLoader=ERROR
-
 log4j.logger.org.apache.spark.storage.BlockManager=ERROR
 log4j.logger.kafka.consumer.ZookeeperConsumerConnector=ERROR
 log4j.logger.kafka.server.HighwaterMarkCheckpoint=ERROR
@@ -20,5 +21,5 @@ log4j.logger.org.apache.hadoop=ERROR
 log4j.logger.kafka=WARN
 log4j.logger.com.datastax.spark=WARN
 log4j.logger.org.apache.spark=WARN
-log4j.logger.com.databricks.apps.core=INFO
-log4j.logger.com.databricks.apps.weather=INFO
+log4j.logger.com.databricks.apps=INFO
+log4j.logger.com.databricks.apps.weather=DEBUG
