+++ timeseries/scala/timeseries-weather/src/main/resources/logback.xml
@@ -12,6 +12,7 @@
     <logger name="org.apache.hadoop" level="ERROR" />
 
     <logger name="com.databricks.apps" level="DEBUG" />
+    <logger name="com.databricks.apps.weather" level="DEBUG" />
     <logger name="com.datastax.driver" level="WARN" />
     <!-- Change to INFO to see spark-cassandra-connector logging -->
     <logger name="com.datastax.spark" level="WARN" />
