+++ timeseries/scala/timeseries-weather/src/main/scala/com/databricks/apps/weather/AggregationActor.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.databricks.apps.core
+package com.databricks.apps.weather
 
 import java.util.concurrent.TimeoutException
