+++ timeseries/scala/timeseries-weather/src/main/scala/com/databricks/apps/weather/WeatherSettings.scala
@@ -15,8 +15,13 @@
  */
 package com.databricks.apps.weather
 
-import com.databricks.apps.core.Settings
-import com.typesafe.config.Config
+import java.io.{File => JFile}
+
+import scala.util.Try
+import com.datastax.driver.core.ConsistencyLevel
+import com.datastax.spark.connector.cql.{NoAuthConf, PasswordAuthConf, AuthConf}
+import com.typesafe.config.{Config, ConfigFactory}
+import org.apache.spark.Logging
 
 /**
  * (Type safe) application settings. First attempts to acquire from the runtime environment.
@@ -38,8 +43,109 @@ import com.typesafe.config.Config
  *
  * @param conf Optional config for test
  */
-final class WeatherSettings(conf: Option[Config] = None) extends Settings(conf) {
+final class WeatherSettings(conf: Option[Config] = None) extends Serializable with Logging {
 
+  val rootConfig = conf match {
+    case Some(c) => c.withFallback(ConfigFactory.load)
+    case _ => ConfigFactory.load
+  }
+
+  protected val spark = rootConfig.getConfig("spark")
+
+  val SparkMaster = withFallback[String](Try(spark.getString("master")),
+    "spark.master") getOrElse "local[*]"
+
+  val SparkCleanerTtl = withFallback[Int](Try(spark.getInt("cleaner.ttl")),
+    "spark.cleaner.ttl") getOrElse (3600*2)
+
+  val SparkStreamingBatchInterval = withFallback[Int](Try(spark.getInt("streaming.batch.interval")),
+    "spark.streaming.batch.interval") getOrElse 1
+
+  protected val cassandra = rootConfig.getConfig("cassandra")
+
+  val CassandraHosts = withFallback[String](Try(cassandra.getString("connection.host")),
+    "spark.cassandra.connection.host") getOrElse "127.0.0.1"
+
+  logInfo(s"Starting up with spark master '$SparkMaster' cassandra hosts '$CassandraHosts'")
+
+  val CassandraAuthUsername: Option[String] = Try(cassandra.getString("auth.username")).toOption
+    .orElse(sys.props.get("spark.cassandra.auth.username"))
+
+  val CassandraAuthPassword: Option[String] = Try(cassandra.getString("auth.password")).toOption
+    .orElse(sys.props.get("spark.cassandra.auth.password"))
+
+  val CassandraAuth: AuthConf = {
+    val credentials = for (
+      username <- CassandraAuthUsername;
+      password <- CassandraAuthPassword
+    ) yield (username, password)
+
+    credentials match {
+      case Some((user, password)) => PasswordAuthConf(user, password)
+      case None                   => NoAuthConf
+    }
+  }
+
+  val CassandraRpcPort = withFallback[Int](Try(cassandra.getInt("connection.rpc.port")),
+    "spark.cassandra.connection.rpc.port") getOrElse 9160
+
+  val CassandraNativePort = withFallback[Int](Try(cassandra.getInt("connection.native.port")),
+    "spark.cassandra.connection.native.port") getOrElse 9042
+
+  /* Tuning */
+
+  val CassandraKeepAlive = withFallback[Int](Try(cassandra.getInt("connection.keep-alive")),
+    "spark.cassandra.connection.keep_alive_ms") getOrElse 1000
+
+  val CassandraRetryCount = withFallback[Int](Try(cassandra.getInt("connection.query.retry-count")),
+    "spark.cassandra.query.retry.count") getOrElse 10
+
+  val CassandraConnectionReconnectDelayMin = withFallback[Int](Try(cassandra.getInt("connection.reconnect-delay.min")),
+    "spark.cassandra.connection.reconnection_delay_ms.min") getOrElse 1000
+
+  val CassandraConnectionReconnectDelayMax = withFallback[Int](Try(cassandra.getInt("reconnect-delay.max")),
+    "spark.cassandra.connection.reconnection_delay_ms.max") getOrElse 60000
+
+  /* Tuning: Reads */
+  val CassandraReadPageRowSize = withFallback[Int](Try(cassandra.getInt("read.page.row.size")),
+    "spark.cassandra.input.page.row.size") getOrElse 1000
+
+  val CassandraReadConsistencyLevel: ConsistencyLevel = ConsistencyLevel.valueOf(
+    withFallback[String](Try(cassandra.getString("read.consistency.level")),
+      "spark.cassandra.input.consistency.level") getOrElse ConsistencyLevel.LOCAL_ONE.name)
+
+  val CassandraReadSplitSize = withFallback[Long](Try(cassandra.getLong("read.split.size")),
+    "spark.cassandra.input.split.size") getOrElse 100000
+
+  /* Tuning: Writes */
+
+  val CassandraWriteParallelismLevel = withFallback[Int](Try(cassandra.getInt("write.concurrent.writes")),
+    "spark.cassandra.output.concurrent.writes") getOrElse 5
+
+  val CassandraWriteBatchSizeBytes = withFallback[Int](Try(cassandra.getInt("write.batch.size.bytes")),
+    "spark.cassandra.output.batch.size.bytes") getOrElse 64 * 1024
+
+  private val CassandraWriteBatchSizeRows = withFallback[String](Try(cassandra.getString("write.batch.size.rows")),
+    "spark.cassandra.output.batch.size.rows") getOrElse "auto"
+
+  val CassandraWriteBatchRowSize: Option[Int] = {
+    val NumberPattern = "([0-9]+)".r
+    CassandraWriteBatchSizeRows match {
+      case "auto"           => None
+      case NumberPattern(x) => Some(x.toInt)
+      case other =>
+        throw new IllegalArgumentException(
+          s"Invalid value for 'cassandra.output.batch.size.rows': $other. Number or 'auto' expected")
+    }
+  }
+
+  val CassandraWriteConsistencyLevel: ConsistencyLevel = ConsistencyLevel.valueOf(
+    withFallback[String](Try(cassandra.getString("write.consistency.level")),
+      "spark.cassandra.output.consistency.level") getOrElse ConsistencyLevel.LOCAL_ONE.name)
+
+  val CassandraDefaultMeasuredInsertsCount: Int = 128
+
+  /* App Specific Settings */
   protected val weather = rootConfig.getConfig("weather")
 
   val AppName = weather.getString("app-name")
@@ -54,20 +160,31 @@ final class WeatherSettings(conf: Option[Config] = None) extends Settings(conf) 
   val CassandraTableSky = weather.getString("cassandra.table.sky")
   val CassandraTableStations = weather.getString("cassandra.table.stations")
 
+  protected val kafka = ConfigFactory.load.getConfig("kafka")
+  val KafkaEncoderFqcn = kafka.getString("encoder.fqcn")
+  val KafkaDecoderFqcn = kafka.getString("decoder.fqcn")
+  val KafkaPartitioner = kafka.getString("partitioner.fqcn")
+  val KafkaBatchSendSize = kafka.getInt("batch.send.size")
   val KafkaGroupId = kafka.getString("group.id")
   val KafkaTopicRaw = kafka.getString("topic.raw")
 
-  val DataLoadPath = weather.getString("data.load.path")
-  val DataFileExtension = weather.getString("data.file.extension")
-  val IngestionData: Set[String] = {
+  val DataLoadPath = weather.getString("data.path")
+  val IngestionData: Set[JFile] = {
     import java.io.{File => JFile}
 
-    val files = new JFile(s"$DataLoadPath").listFiles.collect {
-      case file if file.exists => file.getAbsolutePath.replace("./", "")
+    val files = new JFile(DataLoadPath).listFiles.filter(_.exists).collect {
+      case file if file.exists => file
     }.toSet
 
     log.info(s"Found ${files.size} data files to load.")
     require(files.nonEmpty, s"Unable to find valid data files at $DataLoadPath")
     files
   }
+
+  /** Attempts to acquire from environment, then java system properties. */
+  private def withFallback[T](env: Try[T], key: String): Option[T] = env match {
+    case null  => None
+    case value => value.toOption
+  }
+
 }
